requirements.txt
python-telegram-bot==21.4
openai== 2.6.1
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
gspread==5.12.4
google-auth==2.35.0
